+++ src/main/java/act/aaa/model/MorphiaUserBase.java
@@ -29,29 +29,33 @@ import act.apidoc.SampleDataCategory;
 import act.db.morphia.MorphiaAdaptiveRecord;
 import act.validation.Password;
 import org.osgl.$;
-import org.osgl.aaa.Permission;
-import org.osgl.aaa.Principal;
-import org.osgl.aaa.Privilege;
-import org.osgl.aaa.Role;
+import org.osgl.aaa.*;
+import org.osgl.util.C;
 import org.osgl.util.S;
+import org.osgl.util.StringTokenSet;
 
-import java.util.List;
-import java.util.Set;
+import java.util.*;
+import javax.persistence.Column;
+import javax.persistence.Transient;
 import javax.validation.constraints.NotNull;
 
 public abstract class MorphiaUserBase<T extends MorphiaUserBase> extends MorphiaAdaptiveRecord<T> implements Principal, UserLinked {
 
+    @Transient
+    private transient Map<String, String> _properties = new HashMap<>();
+
     @NotNull
+    @Column(unique = true, nullable = false, updatable = false)
     public String email;
 
     @SampleData.Category(SampleDataCategory.PRIVILEGE)
-    public int privilege;
+    private int privilege;
 
     @SampleData.Category(SampleDataCategory.PERMISSIONS)
-    public String permissions;
+    private String permissions;
 
     @SampleData.Category(SampleDataCategory.ROLES)
-    public String roles;
+    private String roles;
 
     @Password
     private char[] password;
@@ -86,6 +90,60 @@ public abstract class MorphiaUserBase<T extends MorphiaUserBase> extends Morphia
         return AAALookup.permissions(permissions);
     }
 
+    public T grantPermissions(Permission... permissions) {
+        if (permissions.length > 0) {
+            return grantPermissionByNames(stringOf(C.Array.of(permissions)));
+        }
+        return me();
+    }
+
+    public T grantPermissionByNames(String... permissions) {
+        this.permissions = StringTokenSet.merge(this.permissions, permissions);
+        return me();
+    }
+
+    public T grantPermissions(Collection<Permission> permissions) {
+        if (permissions.isEmpty()) {
+            return me();
+        }
+        return grantPermissionByNames(stringOf(permissions));
+    }
+
+    public T grantPermissionByNames(Collection<String> permissions) {
+        if (permissions.isEmpty()) {
+            return me();
+        }
+        this.permissions = StringTokenSet.merge(C.list(permissions).append(this.permissions));
+        return me();
+    }
+
+    public T grantRoles(Role... roles) {
+        if (roles.length > 0) {
+            return grantRoleByNames(stringOf(C.Array.of(roles)));
+        }
+        return me();
+    }
+
+    public T grantRoleByNames(String... roles) {
+        this.roles = StringTokenSet.merge(this.roles, roles);
+        return me();
+    }
+
+    public T grantRoles(Collection<Role> roles) {
+        if (roles.isEmpty()) {
+            return me();
+        }
+        return grantRoleByNames(stringOf(roles));
+    }
+
+    public T grantRoleByNames(Collection<String> roles) {
+        if (roles.isEmpty()) {
+            return me();
+        }
+        this.roles = StringTokenSet.merge(C.list(roles).append(this.roles));
+        return me();
+    }
+
     @Override
     public String getName() {
         return email;
@@ -93,22 +151,26 @@ public abstract class MorphiaUserBase<T extends MorphiaUserBase> extends Morphia
 
     @Override
     public void setProperty(String key, String value) {
-        putValue(key, value);
+        props().put(key, value);
     }
 
     @Override
     public void unsetProperty(String key) {
-        putValue(key, null);
+        props().remove(key);
     }
 
     @Override
     public String getProperty(String key) {
-        return getValue(key);
+        if ("id".equals(key)) {
+            Object o = $.getProperty(this, "id");
+            return S.string(o);
+        }
+        return props().get(key);
     }
 
     @Override
     public Set<String> propertyKeys() {
-        return keySet();
+        return _properties.keySet();
     }
 
     @Override
@@ -116,4 +178,27 @@ public abstract class MorphiaUserBase<T extends MorphiaUserBase> extends Morphia
         return S.blank(email) ? "unknown-user" : email;
     }
 
+    private Map<String, String> props() {
+        if (null == _properties) {
+            _properties = new HashMap<>();
+        }
+        return _properties;
+    }
+
+    protected final T me() {
+        return (T) this;
+    }
+
+    private static String stringOf(Iterable<? extends AAAObject> aaaObjects) {
+        S.Buffer buf = S.buffer();
+        boolean first = true;
+        for (AAAObject obj : aaaObjects) {
+            if (first) {
+                buf.append(StringTokenSet.SEPARATOR);
+                first = false;
+            }
+            buf.append(obj.getName());
+        }
+        return buf.toString();
+    }
 }